Revert "moved artifacts under src so to build starts"
This reverts commit 41fb836a4980f5a447dd9411037bd8f4618b7d24.
