+++ artifacts/pools/ssContracts/SideStaking.sol/SideStaking.dbg.json
