+++ artifacts/pools/ssContracts/SideStaking.sol/SideStaking.json
