+++ artifacts/templates/ERC20TemplateEnterprise.sol/ERC20TemplateEnterprise.dbg.json
