+++ artifacts/templates/ERC20TemplateEnterprise.sol/ERC20TemplateEnterprise.json
