+++ artifacts/templates/ERC721Template.sol/ERC721Template.dbg.json
