+++ src/pools/dispenser/Dispenser.ts
@@ -3,7 +3,7 @@ import { AbiItem } from 'web3-utils'
 import { Contract } from 'web3-eth-contract'
 import { TransactionReceipt } from 'web3-eth'
 import Decimal from 'decimal.js'
-import defaultDispenserAbi from '../../artifacts/pools/dispenser/Dispenser.sol/Dispenser.json'
+import defaultDispenserAbi from '../../../artifacts/pools/dispenser/Dispenser.sol/Dispenser.json'
 import { LoggerInstance as logger, getFairGasPrice } from '../../utils/'
 import { Datatoken } from '../../tokens'
