+++ test/unit/NftFactory.test.ts
@@ -2,15 +2,15 @@ import { assert, expect } from 'chai'
 import { AbiItem } from 'web3-utils/types'
 import { TestContractHandler } from '../TestContractHandler'
 import Web3 from 'web3'
-import ERC721Factory from '../../src/artifacts/ERC721Factory.sol/ERC721Factory.json'
-import ERC721Template from '../../src/artifacts/templates/ERC721Template.sol/ERC721Template.json'
-import SideStaking from '../../src/artifacts/pools/ssContracts/SideStaking.sol/SideStaking.json'
-import Router from '../../src/artifacts/pools/FactoryRouter.sol/FactoryRouter.json'
-import ERC20Template from '../../src/artifacts/templates/ERC20Template.sol/ERC20Template.json'
-import Dispenser from '../../src/artifacts/pools/dispenser/Dispenser.sol/Dispenser.json'
-import FixedRate from '../../src/artifacts/pools/fixedRate/FixedRateExchange.sol/FixedRateExchange.json'
-import OPFCommunityFeeCollector from '../../src/artifacts/communityFee/OPFCommunityFeeCollector.sol/OPFCommunityFeeCollector.json'
-import PoolTemplate from '../../src/artifacts/pools//balancer/BPool.sol/BPool.json'
+import ERC721Factory from '../../artifacts/ERC721Factory.sol/ERC721Factory.json'
+import ERC721Template from '../../artifacts/templates/ERC721Template.sol/ERC721Template.json'
+import SideStaking from '../../artifacts/pools/ssContracts/SideStaking.sol/SideStaking.json'
+import Router from '../../artifacts/pools/FactoryRouter.sol/FactoryRouter.json'
+import ERC20Template from '../../artifacts/templates/ERC20Template.sol/ERC20Template.json'
+import Dispenser from '../../artifacts/pools/dispenser/Dispenser.sol/Dispenser.json'
+import FixedRate from '../../artifacts/pools/fixedRate/FixedRateExchange.sol/FixedRateExchange.json'
+import OPFCommunityFeeCollector from '../../artifacts/communityFee/OPFCommunityFeeCollector.sol/OPFCommunityFeeCollector.json'
+import PoolTemplate from '../../artifacts/pools//balancer/BPool.sol/BPool.json'
 import { NftFactory, NftCreateData, TokenOrder } from '../../src/factories/NFTFactory'
 import {
   FreCreationParams,
