+++ test/unit/pools/dispenser/Dispenser.test.ts
@@ -1,15 +1,15 @@
 import Web3 from 'web3'
 import { AbiItem } from 'web3-utils'
 import { assert, expect } from 'chai'
-import ERC721Factory from '../../../../src/artifacts/ERC721Factory.sol/ERC721Factory.json'
-import ERC721Template from '../../../../src/artifacts/templates/ERC721Template.sol/ERC721Template.json'
-import SideStaking from '../../../../src/artifacts/pools/ssContracts/SideStaking.sol/SideStaking.json'
-import FactoryRouter from '../../../../src/artifacts/pools/FactoryRouter.sol/FactoryRouter.json'
-import ERC20Template from '../../../../src/artifacts/templates/ERC20Template.sol/ERC20Template.json'
-import DispenserTemplate from '../../../../src/artifacts/pools/dispenser/Dispenser.sol/Dispenser.json'
-import FixedRate from '../../../../src/artifacts/pools/fixedRate/FixedRateExchange.sol/FixedRateExchange.json'
-import PoolTemplate from '../../../../src/artifacts/pools/balancer/BPool.sol/BPool.json'
-import OPFCollector from '../../../../src/artifacts/communityFee/OPFCommunityFeeCollector.sol/OPFCommunityFeeCollector.json'
+import ERC721Factory from '../../../../artifacts/ERC721Factory.sol/ERC721Factory.json'
+import ERC721Template from '../../../../artifacts/templates/ERC721Template.sol/ERC721Template.json'
+import SideStaking from '../../../../artifacts/pools/ssContracts/SideStaking.sol/SideStaking.json'
+import FactoryRouter from '../../../../artifacts/pools/FactoryRouter.sol/FactoryRouter.json'
+import ERC20Template from '../../../../artifacts/templates/ERC20Template.sol/ERC20Template.json'
+import DispenserTemplate from '../../../../artifacts/pools/dispenser/Dispenser.sol/Dispenser.json'
+import FixedRate from '../../../../artifacts/pools/fixedRate/FixedRateExchange.sol/FixedRateExchange.json'
+import PoolTemplate from '../../../../artifacts/pools/balancer/BPool.sol/BPool.json'
+import OPFCollector from '../../../../artifacts/communityFee/OPFCommunityFeeCollector.sol/OPFCommunityFeeCollector.json'
 import { NftFactory, NftCreateData } from '../../../../src/factories/'
 import { Datatoken, DispenserParams } from '../../../../src/tokens/'
 import { Dispenser } from '../../../../src/pools/dispenser/'
